Use kaggle-web-client for kaggle_secrets
